Move remaining embedding domain logic into server/embeddings/ sub-package
server/ root keeps only thin wiring: SetupEmbeddingService, callReducePlugin,
projectToCanvas, and Pulse schedule setup methods that couple to s.daemon.
Everything else — clustering, projection, observer, HTTP handlers, Pulse
handlers — lives in the sub-package with explicit dependency injection.
