+++ server/embeddings/observer_test.go
@@ -1,6 +1,6 @@
 //go:build cgo && rustembeddings
 
-package server
+package embeddings
 
 import (
 	"encoding/json"
