+++ server/embeddings_handlers.go
@@ -5,24 +5,16 @@ package server
 import (
 	"context"
 	"encoding/json"
-	"fmt"
-	"net/http"
 	"os"
-	"sort"
-	"strconv"
-	"strings"
-	"sync"
-	"sync/atomic"
 	"time"
 
 	appcfg "github.com/teranos/QNTX/am"
 	"github.com/teranos/QNTX/ats/embeddings/embeddings"
 	"github.com/teranos/QNTX/ats/storage"
-	"github.com/teranos/QNTX/ats/types"
 	"github.com/teranos/QNTX/errors"
+	serverembeddings "github.com/teranos/QNTX/server/embeddings"
 	grpcplugin "github.com/teranos/QNTX/plugin/grpc"
 	"github.com/teranos/QNTX/plugin/grpc/protocol"
-	"go.uber.org/zap"
 )
 
 // SetupEmbeddingService initializes the embedding service if available
@@ -73,20 +65,20 @@ func (s *QNTXServer) SetupEmbeddingService() {
 	s.embeddingStore = embStore
 
 	// Register observer for automatic embedding of attestations with rich text
-	observer := &EmbeddingObserver{
-		embeddingService: embService,
-		embeddingStore:   embStore,
-		richStore:        storage.NewBoundedStore(s.db, nil, s.logger.Named("auto-embed")),
-		logger:           s.logger.Named("auto-embed"),
-		clusterThreshold: float32(appcfg.GetFloat64("embeddings.cluster_threshold")),
-		projectFunc:      s.projectToCanvas,
-	}
+	observer := serverembeddings.NewEmbeddingObserver(
+		embService,
+		embStore,
+		storage.NewBoundedStore(s.db, nil, s.logger.Named("auto-embed")),
+		s.logger.Named("auto-embed"),
+		float32(appcfg.GetFloat64("embeddings.cluster_threshold")),
+		s.projectToCanvas,
+	)
 
 	// Wire semantic watcher matching through the embedding observer.
 	// After embedding, the observer passes the attestation + vector to the watcher
 	// engine — eliminates redundant GenerateEmbedding FFI calls per semantic watcher.
 	if s.watcherEngine != nil {
-		observer.onEmbedded = s.watcherEngine.OnAttestationEmbedded
+		observer.SetOnEmbedded(s.watcherEngine.OnAttestationEmbedded)
 	}
 
 	storage.RegisterObserver(observer)
@@ -102,288 +94,6 @@ func (s *QNTXServer) SetupEmbeddingService() {
 func hasRustEmbeddings() bool {
 	// This function will be overridden by the build tag
 	return true
-}
-
-// EmbeddingObserver automatically embeds attestations that contain rich text.
-// Implements storage.AttestationObserver — called asynchronously in a goroutine
-// by notifyObservers, so errors are logged but don't block attestation creation.
-// Only attestations with non-empty rich string fields (message, description, etc.)
-// trigger embedding; structural-only attestations are silently skipped.
-type EmbeddingObserver struct {
-	embeddingService interface {
-		GenerateEmbedding(text string) (*embeddings.EmbeddingResult, error)
-		SerializeEmbedding(embedding []float32) ([]byte, error)
-		DeserializeEmbedding(data []byte) ([]float32, error)
-		ComputeSimilarity(a, b []float32) (float32, error)
-		GetModelInfo() (*embeddings.ModelInfo, error)
-	}
-	embeddingStore   *storage.EmbeddingStore
-	richStore        *storage.BoundedStore // Reused across calls for 5-min rich field cache
-	logger           *zap.SugaredLogger
-	clusterMu        sync.RWMutex
-	clusterCache     []storage.ClusterCentroid // loaded once, refreshed on re-cluster
-	clusterThreshold float32                   // minimum similarity for cluster assignment
-	projectFunc      func(embeddingID string, embedding []float32)
-
-	// onEmbedded is called after an attestation is successfully embedded and stored.
-	// The watcher engine uses this to run semantic matching with the pre-computed
-	// embedding, eliminating redundant GenerateEmbedding FFI calls.
-	onEmbedded func(as *types.As, embedding []float32)
-
-	// Periodic summary counters (drained by ticker)
-	embedded      atomic.Int64
-	clusterHits   sync.Map // cluster display name (string) → *atomic.Int64
-	clusterNoise  atomic.Int64
-	clusterLabels sync.Map // cluster_id (int) → label (string), refreshed with centroid cache
-}
-
-// InvalidateClusterCache clears cached centroids so the next prediction reloads from DB.
-func (o *EmbeddingObserver) InvalidateClusterCache() {
-	o.clusterMu.Lock()
-	o.clusterCache = nil
-	o.clusterMu.Unlock()
-}
-
-// OnAttestationCreated selectively embeds attestations with rich text content.
-func (o *EmbeddingObserver) OnAttestationCreated(as *types.As) {
-	text := o.extractRichText(as)
-	if text == "" {
-		return
-	}
-
-	// Check if already embedded
-	existing, err := o.embeddingStore.GetBySource("attestation", as.ID)
-	if err != nil {
-		o.logger.Warnw("Failed to check existing embedding",
-			"error", errors.Wrapf(err, "attestation %s", as.ID))
-		return
-	}
-	if existing != nil {
-		return
-	}
-
-	// Generate embedding via Rust FFI (~80ms)
-	result, err := o.embeddingService.GenerateEmbedding(text)
-	if err != nil {
-		o.logger.Warnw("Failed to generate embedding",
-			"error", errors.Wrapf(err, "attestation %s (%d chars)", as.ID, len(text)))
-		return
-	}
-
-	blob, err := o.embeddingService.SerializeEmbedding(result.Embedding)
-	if err != nil {
-		o.logger.Warnw("Failed to serialize embedding",
-			"error", errors.Wrapf(err, "attestation %s (%d dimensions)", as.ID, len(result.Embedding)))
-		return
-	}
-
-	modelInfo, err := o.embeddingService.GetModelInfo()
-	if err != nil {
-		o.logger.Warnw("Failed to get model info",
-			"error", errors.Wrapf(err, "attestation %s", as.ID))
-		return
-	}
-
-	model := &storage.EmbeddingModel{
-		SourceType: "attestation",
-		SourceID:   as.ID,
-		Text:       text,
-		Embedding:  blob,
-		Model:      modelInfo.Name,
-		Dimensions: modelInfo.Dimensions,
-	}
-	if err := o.embeddingStore.Save(model); err != nil {
-		o.logger.Warnw("Failed to save embedding",
-			"error", errors.Wrapf(err, "attestation %s", as.ID))
-		return
-	}
-
-	o.embedded.Add(1)
-
-	o.logger.Debugw("Auto-embedded attestation",
-		"attestation_id", as.ID,
-		"text_length", len(text),
-		"inference_ms", result.InferenceMS)
-
-	// Notify watcher engine with the pre-computed embedding for semantic matching
-	if o.onEmbedded != nil {
-		o.onEmbedded(as, result.Embedding)
-	}
-
-	// Predict cluster assignment for the new embedding
-	o.predictCluster(model.ID, as.ID, result.Embedding)
-
-	// Project to 2D canvas if reduce plugin is available
-	if o.projectFunc != nil {
-		o.projectFunc(model.ID, result.Embedding)
-	}
-}
-
-// predictCluster assigns the embedding to the nearest cluster centroid.
-func (o *EmbeddingObserver) predictCluster(embeddingID, attestationID string, embedding []float32) {
-	// Lazy-load centroids from DB
-	o.clusterMu.RLock()
-	centroids := o.clusterCache
-	o.clusterMu.RUnlock()
-
-	if centroids == nil {
-		loaded, err := o.embeddingStore.GetAllClusterCentroids()
-		if err != nil {
-			o.logger.Warnw("Failed to load cluster centroids",
-				"error", errors.Wrapf(err, "attestation %s", attestationID))
-			return
-		}
-		if len(loaded) == 0 {
-			return // no clusters yet
-		}
-		o.clusterMu.Lock()
-		o.clusterCache = loaded
-		o.clusterMu.Unlock()
-		centroids = loaded
-
-		// Refresh cluster label cache
-		o.refreshClusterLabels()
-	}
-
-	clusterID, prob, err := o.embeddingStore.PredictCluster(
-		embedding,
-		centroids,
-		o.embeddingService.DeserializeEmbedding,
-		o.embeddingService.ComputeSimilarity,
-		o.clusterThreshold,
-	)
-	if err != nil {
-		o.logger.Warnw("Failed to predict cluster",
-			"error", errors.Wrapf(err, "attestation %s", attestationID))
-		return
-	}
-
-	if clusterID == storage.ClusterNoise {
-		o.clusterNoise.Add(1)
-		return // below threshold, stays as noise
-	}
-
-	err = o.embeddingStore.UpdateClusterAssignments([]storage.ClusterAssignment{{
-		ID:          embeddingID,
-		ClusterID:   clusterID,
-		Probability: prob,
-	}})
-	if err != nil {
-		o.logger.Warnw("Failed to save predicted cluster assignment",
-			"error", errors.Wrapf(err, "attestation %s embedding %s cluster %d", attestationID, embeddingID, clusterID))
-		return
-	}
-
-	// Track cluster hit for periodic summary
-	name := o.clusterDisplayName(clusterID)
-	val, _ := o.clusterHits.LoadOrStore(name, &atomic.Int64{})
-	val.(*atomic.Int64).Add(1)
-
-	o.logger.Debugw("Predicted cluster for new embedding",
-		"attestation_id", attestationID,
-		"embedding_id", embeddingID,
-		"cluster_id", clusterID,
-		"similarity", prob)
-}
-
-// extractRichText returns the concatenated rich text fields from an attestation's
-// attributes. Returns empty string if no rich text is found — this is the
-// selective gate that prevents embedding structural-only attestations.
-func (o *EmbeddingObserver) extractRichText(as *types.As) string {
-	if as.Attributes == nil || len(as.Attributes) == 0 {
-		return ""
-	}
-
-	richFields := o.richStore.GetDiscoveredRichFields()
-	if len(richFields) == 0 {
-		return ""
-	}
-
-	return extractRichTextFromAttributes(as.Attributes, richFields)
-}
-
-// clusterDisplayName returns a human-readable name for a cluster ID.
-// Uses the cached label if available, falls back to "cluster:<id>".
-func (o *EmbeddingObserver) clusterDisplayName(clusterID int) string {
-	if label, ok := o.clusterLabels.Load(clusterID); ok {
-		return label.(string)
-	}
-	return "cluster:" + strconv.Itoa(clusterID)
-}
-
-// refreshClusterLabels loads cluster labels from the DB into the label cache.
-func (o *EmbeddingObserver) refreshClusterLabels() {
-	clusters, err := o.embeddingStore.GetActiveClusterIdentities()
-	if err != nil {
-		o.logger.Debugw("Failed to refresh cluster labels", "error", err)
-		return
-	}
-	for _, c := range clusters {
-		if c.Label != nil && *c.Label != "" {
-			o.clusterLabels.Store(c.ID, *c.Label)
-		}
-	}
-}
-
-// DrainEmbeddingCounts atomically reads and resets embedding activity counters.
-// Returns total embedded, cluster assignment breakdown, and noise count.
-func (o *EmbeddingObserver) DrainEmbeddingCounts() (embedded int, clusterCounts []string, noise int) {
-	embedded = int(o.embedded.Swap(0))
-	noise = int(o.clusterNoise.Swap(0))
-
-	var pairs []pairCount
-	o.clusterHits.Range(func(key, value any) bool {
-		count := value.(*atomic.Int64).Swap(0)
-		if count > 0 {
-			pairs = append(pairs, pairCount{Key: key.(string), Count: count})
-		}
-		if count == 0 {
-			o.clusterHits.Delete(key)
-		}
-		return true
-	})
-
-	sort.Slice(pairs, func(i, j int) bool {
-		return pairs[i].Count > pairs[j].Count
-	})
-
-	// Show top 5 clusters
-	limit := 5
-	if len(pairs) < limit {
-		limit = len(pairs)
-	}
-	clusterCounts = make([]string, limit)
-	for i := 0; i < limit; i++ {
-		clusterCounts[i] = formatPairCount(pairs[i])
-	}
-
-	return embedded, clusterCounts, noise
-}
-
-// extractRichTextFromAttributes extracts text from the named rich fields in an
-// attestation's attribute map. Shared by EmbeddingObserver and batch handler.
-func extractRichTextFromAttributes(attrs map[string]interface{}, richFields []string) string {
-	var parts []string
-	for _, field := range richFields {
-		value, exists := attrs[field]
-		if !exists {
-			continue
-		}
-		switch v := value.(type) {
-		case string:
-			if v != "" {
-				parts = append(parts, v)
-			}
-		case []interface{}:
-			for _, item := range v {
-				if str, ok := item.(string); ok && str != "" {
-					parts = append(parts, str)
-				}
-			}
-		}
-	}
-
-	return strings.Join(parts, " ")
 }
 
 // callReducePlugin sends an HTTP request to the reduce plugin via gRPC.
@@ -414,51 +124,6 @@ func (s *QNTXServer) callReducePlugin(ctx context.Context, method, path string, 
 			method, path, resp.StatusCode, string(resp.Body))
 	}
 	return resp.Body, nil
-}
-
-// HandleEmbeddingProject runs configured projection methods on all embeddings.
-// POST /api/embeddings/project
-func (s *QNTXServer) HandleEmbeddingProject(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	if s.embeddingService == nil || s.embeddingStore == nil {
-		http.Error(w, "Embedding service not available", http.StatusServiceUnavailable)
-		return
-	}
-
-	methods := appcfg.GetStringSlice("embeddings.projection_methods")
-	if len(methods) == 0 {
-		methods = []string{"umap"}
-	}
-
-	var params *ProjectionParams
-	if r.Body != nil {
-		var req ProjectionParams
-		if err := json.NewDecoder(r.Body).Decode(&req); err == nil {
-			if req.NNeighbors != nil || req.MinDist != nil || req.Perplexity != nil {
-				params = &req
-			}
-		}
-	}
-
-	startTime := time.Now()
-	results, err := RunAllProjections(r.Context(), methods, s.embeddingStore, s.embeddingService, s.callReducePlugin, s.logger, params)
-	if err != nil {
-		s.logger.Errorw("Projection failed", "methods", methods, "error", err)
-		http.Error(w, fmt.Sprintf("Projection failed: %s", err), http.StatusInternalServerError)
-		return
-	}
-
-	totalMS := float64(time.Since(startTime).Milliseconds())
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(map[string]interface{}{
-		"results":  results,
-		"total_ms": totalMS,
-	})
 }
 
 // projectToCanvas projects a single embedding to 2D via the reduce plugin's /transform
